pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vljepa-research"
version = "0.1.0"
description = "VL-JEPA Research Lab: Implementation of Joint Embedding Predictive Architectures"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    # Core ML
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "timm>=1.0.0",
    "einops>=0.7.0",
    "transformers>=4.40.0",

    # Data
    "datasets>=2.18.0",
    "decord>=0.6.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.9.0",

    # Config & Logging
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "tensorboard>=2.16.0",

    # Utilities
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]
distributed = [
    "submitit>=1.5.0",
]
deployment = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "onnx>=1.15.0",
    "onnxruntime-gpu>=1.17.0",
]
robotics = [
    "mujoco>=3.1.0",
    "gymnasium>=0.29.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
